Add InternalAttributeValue
We're looking to support SDK2. The big idea about how to do that is to make the
encryption logic act on an internal version of the AttributeValue object. This
adds a Translator for moving in between Sdk1 AttributeValues and the
InternalAttributeValue. It also introduces an InternalAttributeValueMarshaller,
which houses some encryption logic. There will be several other PRs related to
moving to InternalAttributeValue.